Parallelise final table creation
After the data is inserted into the log tables we then
run select statements on that data. This can be an expensive
process so we want to run this as soon as the data is loaded into
the log table, without necessarily waiting for all the other tables to
be fully loaded.

We're still going to be bound by the slowest step, but this should
reduce loading time.