Move loading ES Variants out of CH loading.
When the variant table of clickhouse is loaded it started to load the data into ES.
This blocks CH from ingesting the remaining data and means some of the insert
statements timeout resulting in missing data. By running this after all the CH inserts
have completed we should avoid this problem.